requirements.txt
pytorch-lightning
pytorch-lightning-bolts
tensorfn
wandb
ninja
arrayfire
jsonnet
